move address normalization to middleware
